- renamed plugin - added dependencies feature
