+++ README.md
@@ -26,7 +26,7 @@ Run ```check_port.py``` to check that everything is setup correctly.
 ## Usage
 To use this extension you need to collect and inject data into the instance you want to publish.
 
-The data member you need to create is "deadlineData". "deadlineData" is a dictionary containing two items; "job" and "plugin". Each of these has dictionaries as values.
+The data member you need to create is ```deadlineData```. ```deadlineData``` is a dictionary containing three items; ```job```, ```plugin``` and ```order```. ```job``` and ```plugin``` has dictionaries as values. ```order``` is an integer.
 
 **job**
 
@@ -74,6 +74,14 @@ All this data has been filled in by collecting data from the scene and applicati
 ```python
 plugin = {'NukeX': False, 'WriteNode': 'Write1', 'Version': '9.0', 'EnforceRenderOrder': True, 'SceneFile': 'L:\ethel_and_ernest_0001\sequences\q000\q000c010\compositing\publish\q000c010.compositing.v002.nk'}
 ```
+
+**order**
+
+This is an optional value to set. This determines dependencies between jobs. Being a simple integer value, you can set which order the jobs get submitted to the Deadline.
+
+In a publish with instances that has orders like; ```Write1.order = 1``` and ```Write2.order = 2```, ```Write2``` will be dependent on ```Write1``` as ```Write1``` has a lower order value.
+
+This works for multiple dependencies as well. So in the case of; ```Write1.order = 1```, ```Write2.order = 1``` and ```Write3.order = 2```, ```Write3``` will be dependent on ```Write1``` and ```Write2```.
 
 **Event Plugins**
